Refactored 'MenuCache' to 'MenuOptimizer'
