+++ src/Servebolt/Admin/AdminGuiController.php
@@ -9,7 +9,7 @@ use Servebolt\Optimizer\Admin\AcceleratedDomainsImageControl\AcceleratedDomainsI
 use Servebolt\Optimizer\Admin\CachePurgeControl\CachePurgeControl;
 use Servebolt\Optimizer\Admin\PerformanceOptimizer\DatabaseOptimizations;
 use Servebolt\Optimizer\Admin\Prefetching\PrefetchingControl;
-use Servebolt\Optimizer\Admin\PerformanceOptimizer\MenuCacheControl;
+use Servebolt\Optimizer\Admin\PerformanceOptimizer\MenuOptimizerControl;
 use Servebolt\Optimizer\Admin\FullPageCacheControl\FullPageCacheControl;
 use Servebolt\Optimizer\Admin\FullPageCacheControl\CacheTtlControl;
 use Servebolt\Optimizer\Admin\GeneralSettings\GeneralSettings;
@@ -251,7 +251,7 @@ class AdminGuiController
         add_submenu_page('servebolt-wp', __('Performance optimizer', 'servebolt-wp'), __('Performance optimizer', 'servebolt-wp'), 'manage_options', 'servebolt-performance-optimizer', [PerformanceOptimizer::getInstance(), 'render']);
         add_submenu_page(null, null, null, 'manage_options', 'servebolt-performance-optimizer-advanced', [PerformanceOptimizerAdvanced::getInstance(), 'render']);
         add_submenu_page(null, null, null, 'manage_options', 'servebolt-performance-optimizer-database', [DatabaseOptimizations::getInstance(), 'render']);
-        add_submenu_page(null, null, null, 'manage_options', 'servebolt-menu-cache', [MenuCacheControl::getInstance(), 'render']);
+        add_submenu_page(null, null, null, 'manage_options', 'servebolt-menu-cache', [MenuOptimizerControl::getInstance(), 'render']);
 
         // Legacy redirect
         add_submenu_page(null, null, null, 'manage_options', 'servebolt-performance-tools', [$this, 'performanceOptimizerLegacyRedirect']);
