+++ tests/Unit/WpMenuOptimizerTest.php
@@ -3,13 +3,13 @@
 namespace Unit;
 
 use ServeboltWPUnitTestCase;
-use Servebolt\Optimizer\MenuCache\MenuCache;
+use Servebolt\Optimizer\MenuOptimizer\MenuOptimizer;
 
 /**
- * Class WpAssetTest
+ * Class WpMenuOptimizerTest
  * @package Unit
  */
-class WpMenuCacheTest extends ServeboltWPUnitTestCase
+class WpMenuOptimizerTest extends ServeboltWPUnitTestCase
 {
     public function testThatMenuGetsCached(): void
     {
@@ -27,7 +27,7 @@ class WpMenuCacheTest extends ServeboltWPUnitTestCase
         $this->assertNotContains('cache', $menu);
         add_filter('sb_optimizer_is_dev_debug', '__return_true');
         add_filter('sb_optimizer_menu_cache_print_cached_comment', '__return_true');
-        MenuCache::init();
+        MenuOptimizer::init();
         $menu = wp_nav_menu($menuArguments);
         $this->assertContains('cache', $menu);
     }
